added back a bignum variant whose modulus parameters are defined at compile time.
This version is an attempt at the best of both worlds. If your modulus is a witness parameter you can use `runtime_bignum`, else the structs and traits in `lib.nr` can be used.

The methods in `lib.nr` now call the `runtime_bignum.nr` methods when relevant so there is no duplicate algorithm logic